--- CardGameDurak.Logic/AdditionalStrategy.cs
@@ -1,96 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using CardGameDurak.Abstractions;
-
-namespace CardGameDurak.Logic;
-
-/// <summary xml:lang = "ru">
-/// Стратегия, используемая ботом на дополнительной сложности.
-/// </summary>
-public class AdditionalStrategy: IBotStrategy
-{
-    /// <inheritdoc/>
-    public bool TryAttack(
-        IReadOnlyCollection<ICard> handCards,
-        IReadOnlyCollection<ICard> desktopCards,
-        int deckSize,
-        out ICard? resultCard)
-    {
-        resultCard = null!;
-        var simpleCards = handCards.Where(x => !x.IsTrump);
-
-        if (desktopCards.Count == 0)
-            resultCard = simpleCards.MinBy(x => x.Rank) ?? handCards.MinBy(x => x.Rank);
-        if (deckSize > 11)
-            resultCard ??= simpleCards.Where(x => desktopCards.Any(y => x.Rank == y.Rank))
-                                      .MinBy(x => x.Rank);
-        else if (deckSize > 4)
-            resultCard ??= simpleCards.Where(x => desktopCards.Any(y => x.Rank == y.Rank))
-                                      .MinBy(x => x.Rank)
-                        ?? handCards.Where(x => desktopCards.Any(y => x.Rank == y.Rank))
-                                    .Where(x => x.Rank < 10)
-                                    .MinBy(x => x.Rank);
-        else resultCard ??= simpleCards.Where(x => desktopCards.Any(y => x.Rank == y.Rank))
-                                       .MinBy(x => x.Rank)
-                        ?? handCards.Where(x => desktopCards.Any(y => x.Rank == y.Rank))
-                                    .MinBy(x => x.Rank);
-
-
-        return resultCard is not null;
-    }
-
-    /// <inheritdoc/>
-    public bool TryDefence(
-        int ownerId,
-        IReadOnlyCollection<ICard> handCards,
-        IReadOnlyCollection<ICard> desktopCards,
-        int deckSize,
-        out ICard? resultCard,
-        out ICard closedCard)
-    {
-        if (desktopCards is null || desktopCards.Count == 0)
-            throw new ArgumentNullException(nameof(desktopCards));
-
-        var needClosed = desktopCards.Where(x => x.Owner!.Id != ownerId)
-                                     .First(x => !x.IsCloseOnDesktop);
-        closedCard = needClosed;
-        var simpleCards = handCards.Where(x => !x.IsTrump);
-        var trumpCards = handCards.Where(x => x.IsTrump);
-        resultCard = simpleCards.Where(x => needClosed.Rank < x.Rank)
-                                .Where(x => needClosed.Suit == x.Suit)
-                                .MinBy(x => x.Rank);
-        if (deckSize > 11)
-        {
-            resultCard ??= handCards.Where(x => x.IsTrump)
-                                    .Where(needClosed => !needClosed.IsTrump)
-                                    .Where(x => x.Rank <= 10)
-                                    .MinBy(x => x.Rank);
-        }
-        else if (deckSize > 4)
-        {
-            resultCard ??= handCards.Where(x => x.IsTrump)
-                                    .Where(x => !needClosed.IsTrump)
-                                    .MinBy(x => x.Rank);
-            resultCard ??= handCards.Where(x => x.IsTrump)
-                                    .Where(x => needClosed.IsTrump)
-                                    .Where(x => x.Rank > needClosed.Rank)
-                                    .Where(x => needClosed.Rank <= 10)
-                                    .MinBy(x => x.Rank);
-        }
-        else
-        {
-            resultCard ??= handCards.Where(x => x.IsTrump)
-                                    .Where(x => !needClosed.IsTrump)
-                                    .MinBy(x => x.Rank) 
-                       ??  handCards.Where(x => x.IsTrump)
-                                    .Where(x => needClosed.IsTrump)
-                                    .Where(x => x.Rank > needClosed.Rank)
-                                    .MinBy(x => x.Rank);
-        }
-
-        return resultCard is not null;
-    }
-}
